html: layout: Split out common helpers
